started implementing lod2 tab MVP, debugging required
